+++ apps/docs/src/app/core/component-docs/datetime-picker/examples/datetime-form-example/datetime-form-example.component.ts
@@ -1,4 +1,4 @@
-import { Component } from '@angular/core';
+import { Component, } from '@angular/core';
 import { FormControl, FormGroup } from '@angular/forms';
 import { FdDatetime } from '@fundamental-ngx/core';
 
@@ -13,8 +13,10 @@ import { FdDatetime } from '@fundamental-ngx/core';
     templateUrl: './datetime-form-example.component.html'
 })
 export class DatetimeFormExampleComponent {
+
     customForm = new FormGroup({
         date: new FormControl(FdDatetime.getToday()),
         disabledDate: new FormControl({ value: FdDatetime.getToday(), disabled: true })
     });
+
 }
