+++ apps/docs/src/app/core/component-docs/icon/examples/icon-example.component.scss
