+++ apps/docs/src/app/core/component-docs/icon/examples/icon-example.component.ts
@@ -3,7 +3,7 @@ import { Component } from '@angular/core';
 @Component({
     selector: 'fd-icon-example',
     templateUrl: './icon-example.component.html',
-    styleUrls: ['icon-docs.component.scss']
+    styleUrls: ['icon-example.component.scss']
 })
 export class IconExampleComponent {
