+++ apps/docs/src/app/core/component-docs/inline-help/examples/inline-help-example.component.scss
@@ -1,0 +1,4 @@
+.fd-inline-help__content {
+    visibility: visible;
+    opacity: 1;
+}
