+++ apps/docs/src/app/core/component-docs/multi-input/examples/multi-input-form-example/multi-input-form-example.component.html
@@ -1,22 +1,31 @@
 <form [formGroup]="customForm">
     <div fd-form-set class="flex-form-set">
         <div fd-form-item>
-            <fd-multi-input formControlName="selectedValues" [dropdownValues]="['Apple', 'Banana', 'Pineapple', 'Tomato']"
-                [placeholder]="'Search Here...'">
+            <fd-multi-input
+                formControlName="selectedValues"
+                [dropdownValues]="['Apple', 'Banana', 'Pineapple', 'Tomato']"
+                [placeholder]="'Search Here...'"
+            >
             </fd-multi-input>
         </div>
         <div fd-form-item>
-            <fd-multi-input formControlName="disabledSelectedValues" [dropdownValues]="['Apple', 'Banana', 'Pineapple', 'Tomato']"
-                    [placeholder]="'Search Here...'">
+            <fd-multi-input
+                formControlName="disabledSelectedValues"
+                [dropdownValues]="['Apple', 'Banana', 'Pineapple', 'Tomato']"
+                [disabled]="'true'"
+                [placeholder]="'Search Here...'"
+            >
             </fd-multi-input>
         </div>
     </div>
 </form>
 
-Selected Values: <span *ngIf="customForm.controls.selectedValues.value">
+Selected Values:
+<span *ngIf="customForm.controls.selectedValues.value">
     {{ customForm.controls.selectedValues.value | json }}
 </span>
 
-Selected Values: <span *ngIf="customForm.controls.disabledSelectedValues.value">
+Selected Values:
+<span *ngIf="customForm.controls.disabledSelectedValues.value">
     {{ customForm.controls.disabledSelectedValues.value | json }}
 </span>
