+++ apps/docs/src/app/core/component-docs/split-button/examples/split-button-options-example.component.ts
@@ -1,0 +1,17 @@
+import { Component } from '@angular/core';
+
+@Component({
+    selector: 'fd-split-button-options-example',
+    templateUrl: './split-button-options-example.component.html',
+    styleUrls: ['./split-button-examples.component.scss']
+})
+export class ButtonSplitOptionsExampleComponent {
+    itemClicked() {
+        alert('Item Clicked!');
+    }
+
+    primaryButtonClicked() {
+        alert('Primary Button Clicked!');
+    }
+
+}
