+++ apps/docs/src/app/core/component-docs/split-button/examples/split-button-programmatical-example.component.ts
@@ -1,0 +1,19 @@
+import { Component } from '@angular/core';
+
+@Component({
+    selector: 'fd-split-button-programmatical-example',
+    templateUrl: './split-button-programmatical-example.component.html',
+    styleUrls: ['./split-button-examples.component.scss']
+})
+export class ButtonSplitProgrammaticalExampleComponent {
+    isOpen: boolean = false;
+
+    itemClicked() {
+        alert('Item Clicked!');
+    }
+
+    primaryButtonClicked() {
+        alert('Primary Button Clicked!');
+    }
+
+}
