+++ apps/docs/src/app/core/component-docs/split-button/split-button-docs.component.ts
@@ -2,8 +2,12 @@ import { Component, OnInit, ViewChildren, ElementRef, QueryList } from '@angular
 import { Schema } from '../../../schema/models/schema.model';
 import { SchemaFactoryService } from '../../../schema/services/schema-factory/schema-factory.service';
 
+import * as buttonSplitOptionsExampleTs from '!raw-loader!./examples/split-button-options-example.component.ts';
+import * as buttonSplitProgrammaticalyExampleTs from '!raw-loader!./examples/split-button-programmatical-example.component.ts';
+import * as buttonSplitTemplateExampleTs from '!raw-loader!./examples/split-button-template-example.component.ts';
+import * as buttonTypesExampleTs from '!raw-loader!./examples/split-button-types-example.component.ts';
+import * as buttonSplitIconsTs from '!raw-loader!./examples/split-button-icons-example.component.ts';
 import * as buttonTypesExample from '!raw-loader!./examples/split-button-types-example.component.html';
-import * as splitButtonTscode from '!raw-loader!./examples/split-button-examples.component.ts';
 import * as splitButtonScsscode from '!raw-loader!./examples/split-button-examples.component.scss';
 import * as buttonSplitProgrammaticalyExample from '!raw-loader!./examples/split-button-programmatical-example.component.html';
 import * as buttonSplitOptionsExample from '!raw-loader!./examples/split-button-options-example.component.html';
@@ -36,10 +40,9 @@ export class SplitButtonDocsComponent implements OnInit {
             code: buttonTypesExample,
             fileName: 'split-button-types-example',
             secondFile: 'split-button-examples',
-            typescriptFileCode: splitButtonTscode,
+            typescriptFileCode: buttonTypesExampleTs,
             component: 'ButtonSplitTypesExampleComponent',
             scssFileCode: splitButtonScsscode,
-            addonExport: 'itemClicked() {alert("Item Clicked!");}   primaryButtonClicked() {alert("Primary Button Clicked!");}'
         }
     ];
 
@@ -49,10 +52,9 @@ export class SplitButtonDocsComponent implements OnInit {
             code: buttonSplitProgrammaticalyExample,
             fileName: 'split-button-programmatical-example',
             secondFile: 'split-button-examples',
-            typescriptFileCode: splitButtonTscode,
+            typescriptFileCode: buttonSplitProgrammaticalyExampleTs,
             component: 'ButtonSplitProgrammaticalExampleComponent',
             scssFileCode: splitButtonScsscode,
-            addonExport: 'itemClicked() {alert("Item Clicked!");}   primaryButtonClicked() {alert("Primary Button Clicked!");}'
         }
     ];
 
@@ -62,10 +64,9 @@ export class SplitButtonDocsComponent implements OnInit {
             code: buttonSplitTemplateExample,
             fileName: 'split-button-template-example',
             secondFile: 'split-button-examples',
-            typescriptFileCode: splitButtonTscode,
+            typescriptFileCode: buttonSplitTemplateExampleTs,
             component: 'ButtonSplitTemplateExampleComponent',
             scssFileCode: splitButtonScsscode,
-            addonExport: 'itemClicked() {alert("Item Clicked!");}   primaryButtonClicked() {alert("Primary Button Clicked!");}'
         }
     ];
 
@@ -75,10 +76,9 @@ export class SplitButtonDocsComponent implements OnInit {
             code: buttonSplitOptionsExample,
             fileName: 'split-button-options-example',
             secondFile: 'split-button-examples',
-            typescriptFileCode: splitButtonTscode,
+            typescriptFileCode: buttonSplitOptionsExampleTs,
             component: 'ButtonSplitOptionsExampleComponent',
             scssFileCode: splitButtonScsscode,
-            addonExport: 'itemClicked() {alert("Item Clicked!");}   primaryButtonClicked() {alert("Primary Button Clicked!");}'
         }
     ];
 
@@ -88,10 +88,9 @@ export class SplitButtonDocsComponent implements OnInit {
             code: buttonSplitIcons,
             fileName: 'split-button-icons-example',
             secondFile: 'split-button-examples',
-            typescriptFileCode: splitButtonTscode,
+            typescriptFileCode: buttonSplitIconsTs,
             component: 'ButtonSplitTypesIconsComponent',
             scssFileCode: splitButtonScsscode,
-            addonExport: 'itemClicked() {alert("Item Clicked!");}   primaryButtonClicked() {alert("Primary Button Clicked!");}'
         }
     ];
