+++ apps/docs/src/app/core/component-docs/time-picker/examples/time-picker-12-example.component.ts
@@ -1,0 +1,11 @@
+import { Component } from '@angular/core';
+
+@Component({
+    selector: 'fd-time-picker-12-example',
+    templateUrl: './time-picker-12-example.component.html'
+})
+export class TimePicker12ExampleComponent {
+
+    timeMeridianObject = { hour: 12, minute: 0, second: 0 };
+
+}
