+++ apps/docs/src/app/core/component-docs/time-picker/examples/time-picker-allow-null-example.component.ts
@@ -1,0 +1,11 @@
+import { Component } from '@angular/core';
+
+@Component({
+    selector: 'fd-time-picker-allow-null-example',
+    templateUrl: './time-picker-allow-null-example.component.html'
+})
+export class TimePickerAllowNullExampleComponent {
+
+    timeObject = { hour: 12, minute: 0, second: 0 };
+
+}
