+++ apps/docs/src/app/core/component-docs/time-picker/examples/time-picker-compact-example.component.ts
@@ -1,0 +1,11 @@
+import { Component } from '@angular/core';
+
+@Component({
+    selector: 'fd-time-picker-compact-example',
+    templateUrl: './time-picker-compact-example.component.html'
+})
+export class TimePickerCompactExampleComponent {
+
+    timeObject = { hour: 12, minute: 0, second: 0 };
+
+}
