+++ apps/docs/src/app/core/component-docs/time-picker/examples/time-picker-example.component.ts
@@ -1,0 +1,11 @@
+import { Component } from '@angular/core';
+
+@Component({
+    selector: 'fd-time-picker-example',
+    templateUrl: './time-picker-example.component.html'
+})
+export class TimePickerExampleComponent {
+
+    timeObject = { hour: 12, minute: 0, second: 0 };
+
+}
