+++ apps/docs/src/app/documentation/core-helpers/code-example/example-file.ts
@@ -16,6 +16,5 @@ export interface ExampleFile {
     thirdFile?: string;
     component?: string;
     fileName?: string;
-    addonExport?: string;
     appModuleAddon?: string;
 }
